arg-classifier/requirements.txt
biopython==1.81
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1
matplotlib==3.7.2
seaborn==0.12.2
